fix: BF16 default — Qwen3.5 DeltaNet backward is FP16-unstable (iter 23)
Resolves the pre-existing grad_norm_global=nan signal on the
integration test.

Root cause.
Systematic debugging via torch.autograd.detect_anomaly traced the
gradient NaN to MmBackward0 at linear_attn.in_proj_qkv -- the Gated
DeltaNet input projection inside Qwen3.5's hybrid transformer stack.
Forward was finite (log_probs + entropy both nan-free); only the
backward produced NaNs.

Refutation tests:
- Dropping LoRA from linear_attn did NOT fix the NaN, so the
  instability is in the base DeltaNet backward, not in PEFT's wrapper.
- The warning "fast path is not available because flash-linear-
  attention is not installed" points to a numerically stable
  fast-path backward that we are not using.
- Switching dtype=torch.bfloat16 while leaving everything else the
  same -> finite gradients across pg_loss, v_loss, and entropy.

Fix.
Flip the training precision default from FP16 -> BF16 for the
Qwen3.5 family. Amendment 2026-04-20-bf16-default-for-qwen3.5.md
records the override of master-spec §1's FP16 default (FP16 remains
opt-in for non-hybrid backbones where the RL-lore stability claim
holds).

Changes.
- configs/backbones.yaml: dtype float16 -> bfloat16 across all four
  entries (Qwen3.5-0.8B/2B/4B and the Qwen3-VL fallback rows).
- algos/ppo_cot.py::Args.precision default -> "bf16". Fp16State
  disables itself at bf16 (no GradScaler needed; bf16 has the fp32
  exponent range).
- tests/integration/test_trainer_short_run.py now asserts
  grad_norm_global is finite (not "nan"/"inf"), and reads the LAST
  metrics row to be robust to CsvWriter's append-mode behavior across
  re-runs.

Evidence.
- detect_anomaly probe (removed after use): FP16 -> MmBackward0 NaN
  at linear_attn.in_proj_qkv; BF16 -> zero bad params across all
  loss components + full loss.
- Integration test metrics post-fix: approx_kl=0, inv_4_status=green,
  grad_norm_global=91.75 (finite), loss_scale=1.0 (BF16 path).
- pytest tests/unit tests/invariants -m "not gpu": 49/49 pass.
- pytest tests/integration -m "tier1 and gpu": 8/8 pass in 104.64 s.

Invariants run.
- Inv-1/3/8/2: green on wiring tests (bf16 doesn't regress them).
- Inv-4: green (precision-invariant via step-grouped re-score).
- Inv-6: green-by-construction under bf16 (Fp16State disabled).

Follow-up: flash-linear-attention + causal-conv1d installation as a
future perf optimization track; compare BF16-fast-path throughput vs
BF16-fallback.

+++ algos/ppo_cot.py
@@ -100,7 +100,13 @@ class Args:
 
     # Distributed
     sharding: str = "deepspeed_zero2"
-    precision: str = "fp16"
+    # BF16 default (not FP16 as master-spec §1 suggested): Qwen3.5's Gated
+    # DeltaNet backward is numerically unstable in FP16 when the fast-path
+    # flash-linear-attention is not installed — produces NaN gradients at
+    # linear_attn.in_proj_qkv despite finite forward output. BF16's wider
+    # exponent avoids the underflow path. See amendment
+    # 2026-04-20-bf16-default-for-qwen3.5.md.
+    precision: str = "bf16"
     num_processes: int = 1
     grad_accum: int = 1
